fix: migrate every iconText caller, roll the frontend ConfigMap, fix iOS build
Three CI/review failures from the previous push.

**iOS build.** MessageDto.channelSeq is Int64?, so `?? Int.max` did not compile.
Use Int64.max. `swiftc -parse` accepted it because the mismatch is in type
checking, not syntax — the PR flagged that parse-only was not enough, and it was
this that slipped through.

**Every iconText caller, not 17 of them.** The previous sweep matched `<Button`
only, and 52 files import the primitive as `UiButton`, so it missed 20 sites.
The dev-time guard would have thrown on common surfaces (chat selection
toolbar, connector manager, remote workspace dialog), which is what the review
caught. All are migrated now, and several were rendering the wrong label all
along rather than merely a duplicated one:

- the chat selection toolbar's Forward button carried `action="send"`, so it
  rendered "Send"
- the discussion back control carried `action="start"`, so it rendered "Start"
- the remote-workspace diff close button carried `action="preview"`
- the working-tree / staged diff scope buttons shared `close` and `preview`
- the mic toggle rendered "Turn off"/"Turn on" over a dropped "Mute"/"Unmute"

New registry keys (forward, manage, mute, unmute, diffWorking, diffStaged), all
inside the existing slot budget. PermissionCard's diff spinner moves to the
primitive's loading prop.

To stop this recurring, action-label-callers.test.ts scans the source for the
pattern across both import names, so a caller cannot hide behind an alias or in
a branch that tests never render.

**Helm rollout.** The pod-template checksum hashed only the gateway host:port,
so `helm upgrade` left running pods serving the previous CSP. The nginx config
and the header snippet move into named templates that both the ConfigMap and
the checksum consume, so any policy edit now rolls the pods. Verified by
diffing the rendered checksum across a CSP change.

**Security check.** frontend-security.test.mjs asserted a CSP inside
nginx.conf, which no longer holds it. It now checks the snippet and the chart,
asserts the two copies have not drifted, and asserts every nginx location that
sets add_header still includes the shared file.

+++ apps/ios/Sources/Models/MessageTree.swift
@@ -71,8 +71,8 @@ enum MessageTree {
     /// is the tiebreak so rows with equal (or absent) sequence numbers keep a
     /// stable position instead of reshuffling between renders.
     static func precedesInThread(_ lhs: MessageDto, _ rhs: MessageDto) -> Bool {
-        let left = lhs.channelSeq ?? Int.max
-        let right = rhs.channelSeq ?? Int.max
+        let left = lhs.channelSeq ?? Int64.max
+        let right = rhs.channelSeq ?? Int64.max
         return left == right ? lhs.msgId < rhs.msgId : left < right
     }
